# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LifeLoom API",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organs": {
            "get": {
                "tags": [
                    "Órgãos"
                ],
                "summary": "Listar todos os órgãos",
                "operationId": "c4bd32fce6c92d05f87d336dc950aabc",
                "responses": {
                    "200": {
                        "description": "Lista de órgãos retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Órgãos listados com sucesso."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "organ_type": {
                                                        "type": "string",
                                                        "example": "Coração"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Available"
                                                    },
                                                    "expiration_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-12-31"
                                                    },
                                                    "distance_limit": {
                                                        "type": "integer",
                                                        "example": 300
                                                    },
                                                    "hospital_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "donor_id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "recipient_id": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "matched_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-01T10:00:00Z"
                                                    },
                                                    "completed_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-02T15:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao processar a solicitação."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Órgãos"
                ],
                "summary": "Registrar um novo órgão",
                "operationId": "f6317918726b853cb18755ba5df9678f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "organ_type",
                                    "status",
                                    "expiration_date",
                                    "distance_limit",
                                    "hospital_id"
                                ],
                                "properties": {
                                    "organ_type": {
                                        "type": "string",
                                        "example": "Coração"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Available"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "distance_limit": {
                                        "type": "integer",
                                        "example": 300
                                    },
                                    "hospital_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "donor_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "recipient_id": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Órgão criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Órgão registrado com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "organ_type": {
                                                    "type": "string",
                                                    "example": "Coração"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "Available"
                                                },
                                                "expiration_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-31"
                                                },
                                                "distance_limit": {
                                                    "type": "integer",
                                                    "example": 300
                                                },
                                                "hospital_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "donor_id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "recipient_id": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "matched_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-01T10:00:00Z"
                                                },
                                                "completed_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-02T15:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao processar a solicitação."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}